+++ marshmallow/exceptions.py
@@ -40,8 +40,18 @@ class ForcedError(_WrappingException):
 
 
 class ValidationError(MarshmallowError):
-    """Raised when validation fails on a field."""
-    pass
+    """Raised when validation fails on a field.
+
+    :param message: An error message, list of error messages, or dict of
+        error messages.
+    """
+
+    def __init__(self, message):
+        if not isinstance(message, dict) and not isinstance(message, list):
+            messages = [message]
+        else:
+            messages = message
+        self.messages = messages
 
 
 class RegistryError(ForcedError, NameError):