+++ marshmallow/validate.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 """Validation functions for various types of data."""
 import re
+from marshmallow.exceptions import ValidationError
 
 
 URL_REGEX = re.compile(
@@ -26,20 +27,20 @@ RELATIVE_URL_REGEX = re.compile(
     r'(?:/?|[/?]\S+)$', re.IGNORECASE)  # host is optional, allow for relative URLs
 
 
-def url(value, relative=False):
+def url(value, relative, error):
     """Validate a URL.
 
     :param string value: The URL to validate
     :param bool relative: Whether to allow relative URLs.
     :returns: The URL if valid.
-    :raises: ValueError if url is invalid.
+    :raises: ValidationError if url is invalid.
     """
     regex = RELATIVE_URL_REGEX if relative else URL_REGEX
-    if not regex.search(value):
+    if value and not regex.search(value):
         message = u'"{0}" is not a valid URL'.format(value)
         if regex.search('http://' + value):
             message += u'. Did you mean: "http://{0}"?'.format(value)
-        raise ValueError(message)
+        raise ValidationError(error or message)
     return value
 
 USER_REGEX = re.compile(
@@ -56,21 +57,24 @@ DOMAIN_REGEX = re.compile(
 DOMAIN_WHITELIST = ("localhost", )
 
 
-def email(value):
+def email(value, error=None):
     """Validate an email address.
 
+    :param str error: Error message to show.
     :param string value: The email address to validate.
     :returns: The email address if valid.
-    :raises: ValueError if email is invalid
+    :raises: ValidationError if email is invalid
     """
-    error_message = '"{0}" is not a valid email address.'.format(value)
+    if value is None:
+        return None
+    error_message = error or '"{0}" is not a valid email address.'.format(value)
     if not value or '@' not in value:
-        raise ValueError(error_message)
+        raise ValidationError(error_message)
 
     user_part, domain_part = value.rsplit('@', 1)
 
     if not USER_REGEX.match(user_part):
-        raise ValueError(error_message)
+        raise ValidationError(error_message)
 
     if (domain_part not in DOMAIN_WHITELIST and
             not DOMAIN_REGEX.match(domain_part)):
@@ -78,8 +82,8 @@ def email(value):
         try:
             domain_part = domain_part.encode('idna').decode('ascii')
             if not DOMAIN_REGEX.match(domain_part):
-                raise ValueError(error_message)
+                raise ValidationError(error_message)
         except UnicodeError:
             pass
-        raise ValueError(error_message)
+        raise ValidationError(error_message)
     return value